:hammer: improve structure
